bigint > int; validators updates; removed some not needed "contextData"; db structure updates; fixes in place.controller on transaction usage; lang vars usage update: shared.error > shared.validation; some static fn in repository.abstract; and category 90% - still requires images related updates

+++ src/features/account/account-recovery.entity.ts
@@ -20,10 +20,10 @@ export default class AccountRecoveryEntity {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = false;
 
-	@PrimaryGeneratedColumn({ type: 'bigint', unsigned: false })
+	@PrimaryGeneratedColumn({ type: 'int', unsigned: false })
 	id!: number;
 
-	@Column('bigint', { unsigned: false, nullable: false })
+	@Column('int', { unsigned: false, nullable: false })
 	@Index('IDX_account_recovery_user_id')
 	user_id!: number;
