+++ src/features/invoice/invoice.entity.ts
@@ -84,7 +84,7 @@ export default class InvoiceEntity extends EntityAbstract {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = true;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	@Index('IDX_invoice_order_id')
 	order_id!: number;
