+++ src/features/order/order-product.entity.ts
@@ -15,11 +15,11 @@ export default class OrderProductEntity extends EntityAbstract {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = true;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	@Index('IDX_order_product_order_id')
 	order_id!: number;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	@Index('IDX_order_product_product_id')
 	product_id!: number;
