+++ src/features/permission/permission.entity.ts
@@ -18,7 +18,7 @@ export default class PermissionEntity {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = true;
 
-	@PrimaryGeneratedColumn({ type: 'bigint', unsigned: false })
+	@PrimaryGeneratedColumn({ type: 'int', unsigned: false })
 	id!: number;
 
 	@Column('varchar', { nullable: false })
