+++ src/features/place/place-content.entity.ts
@@ -16,7 +16,7 @@ export default class PlaceContentEntity extends EntityAbstract {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = true;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	place_id!: number;
 
 	@Column('varchar', {
