+++ src/features/product/product-attribute.entity.ts
@@ -21,14 +21,14 @@ export default class ProductAttributeEntity extends EntityAbstract {
 	static readonly NAME: string = ENTITY_TABLE_NAME;
 	static readonly HAS_CACHE: boolean = true;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	@Index('IDX_product_attribute_product_id')
 	product_id!: number;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	attribute_label_id!: number;
 
-	@Column('bigint', { nullable: false })
+	@Column('int', { nullable: false })
 	attribute_value_id!: number;
 
 	// RELATIONS
